+++ crates/git/Cargo.toml
@@ -13,6 +13,7 @@ path = "src/git.rs"
 
 [dependencies]
 anyhow.workspace = true
+async-trait.workspace = true
 clock.workspace = true
 collections.workspace = true
 git2.workspace = true
