+++ crates/git/src/hosting_providers.rs
@@ -1,0 +1,13 @@
+mod bitbucket;
+mod codeberg;
+mod gitee;
+mod github;
+mod gitlab;
+mod sourcehut;
+
+pub use bitbucket::*;
+pub use codeberg::*;
+pub use gitee::*;
+pub use github::*;
+pub use gitlab::*;
+pub use sourcehut::*;
